% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_sandbox_mode}
\alias{is_sandbox_mode}
\title{Check whether sandbox mode is enabled globally}
\usage{
is_sandbox_mode()
}
\value{
logical; \code{TRUE} if sandbox mode is enabled globally
}
\description{
Determines the default value of the \code{sandbox} argument of \code{\link[=make_request]{make_request()}}.
Sandbox mode is enabled when the \code{csiapps.sandbox} R option is \code{TRUE}, or,
if the option is unset, when the \code{CSIAPPS_ENV} environment variable equals
\code{"sandbox"}.
}
\examples{
is_sandbox_mode()

options(csiapps.sandbox = TRUE)
is_sandbox_mode()
options(csiapps.sandbox = NULL)
}
\seealso{
\link{csiapps-sandbox} for an overview of sandbox mode
}
