% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_institute}
\alias{set_institute}
\title{Set the target institute for API calls}
\usage{
set_institute(institute = "csipacific")
}
\arguments{
\item{institute}{One of "csipacific", "csiontario", or "csiatlantic"}
}
\description{
Set the target institute for API calls
}
\examples{
set_institute("csiontario")
}
