% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csiapps-package.R
\docType{package}
\name{csiapps-package}
\alias{csiapps}
\alias{csiapps-package}
\title{csiapps: Helper functions for CSIAPPS}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Helper functions and utilities for CSI data warehouse ingestion and shiny web applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CSIOntario/csiapps}
  \item \url{https://csiontario.github.io/csiapps/}
  \item Report bugs at \url{https://github.com/CSIOntario/csiapps/issues}
}

}
\author{
\strong{Maintainer}: David Awosoga \email{dawosoga@csiontario.ca} (\href{https://orcid.org/0000-0002-2741-5770}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kyu Min Shim (\href{https://orcid.org/0009-0006-4161-9813}{ORCID})
}

}
\keyword{internal}
